Moved classes around to clean things up
