+++ TeamCode/src/main/java/com/edinaftc/opmodes/roadrunner/tuner/TrackWidthTuner.java
@@ -1,4 +1,4 @@
-package com.edinaftc.opmodes.roadrunner;
+package com.edinaftc.opmodes.roadrunner.tuner;
 
 import com.acmerobotics.dashboard.FtcDashboard;
 import com.acmerobotics.dashboard.config.Config;
